git-rj: Tool in python to support git usage
The repository will have multiple submodules. The 'git-rj' tool will help with initializing (after checkout), and switching to the default branch for the submodules. It requires Python 3.9 and works with Git 2.13 and later.

To get help, issue the command

  git rj help

To get help for a specific command, use the -h option

  git rj status -h

See the README.gitrj.md file for detailed information about how to use this script.

Use the bootstrapper script "git-rj", which will start the correct version of Python on the actual script content "git-rj.py". This is for compatibility with both Windows and Linux.

Tested on:
- Windows 10 v2004 with Git for Windows 2.13 and Python 3.9.0.
- Ubuntu 18.04.5 with Git v2.17.1 and Python 3.6.9

Issue: DOTNET-239