git-rj: Fix crashes when on a branch that has no commits
When on a branch that has no commits, the command `git rev-parse --abbrev-ref HEAD` returns an error. This needed to be replaced with `git branch --show-current` which works on GIT 2.22 and later.

Then when getting the current commit for that branch, we'll get "None", which is correct, as there is no commit. In this case, we translate that to a default hash of "0000000000000000000000000000000000000000". That at least allows the output to be processed without big changes.

Issue: DOTNET-257