Git-RJ: Print reason why GIT fails
In the case of pulling a repository on a branch, the output for the error would be 'None'. Now print the actual reason for the failure.

The reason for the failure is that the __str__ method for the GitError exception would return the name of the error type, which is always 'None' in the case of a string. Instead, now check if the error type is not defined, then we print the string message.