Update git-rj to build based on a json file
In the previous commit, the build scripts were removed, and we no longer rely on the custom RjBuild tool written for .NET 4.x as it doesn't work with the newer .NET Core and the newer .NET SDK project format.

It is now reduced significantly in complexity automating building also on Linux, using Python (quicker to write and easier to maintain than the C# tool). We do lose some functionality:
- Lost test capabilities for resource locking. Test cases shouldn't use global resources, or should manage that themselves if they're to run in parallel.
- No incremental testing based on git GIT commit identifier
- Release checks based on GIT are no longer done
- NuGet versions based on current commit
- Output put into a 'distribute' folder. It's spread out, and no "nice" reporting like before

What we now gain is:
- Can use cross-platform builds. The .gitrjbuild configuration file defines what is built based on the current platform

Some time after this commit, we need to MSBuild to bring back some of the important features that check for mistakes when doing a release (such as checking GIT labels, and automatic versioning based on current GIT commits). The idea is that this can be put in the .csproj files directly and integrate better with MSBuild so it doesn't break so often.

Issue: DOTNET-343, DOTNET-346