Use .NET SDK Project Format
Convert all projects to compile with the `dotnet` command. It will compile also on Linux (but to do so, there is a solution provided which doesn't include Windows specific projects that would otherwise result in the error NETSDK1100).

The SHFB project file is updated to obtain the documentation from the latest .NET 4.x release build output.

Issue: DOTNET-343