SignTool: Execute SignTool on binaries when doing a release build
All projects have added the NuGet package RJCP.MSBuildTasks that contains a target to sign the assembly after compilation. By default, when building in any release, the targets are not signed.

The signing occurs when the command `git rj build --release` is given, which takes `signcert.crt` and obtains its thumbprint, which is then used to give to signtool to sign (so the private key must be committed into the Windows certificate store). The command `--release` calls `dotnet build -c Release /p:X509SigningCert=signcert.crt /p:X509TimeStampUri=http://timestamp.digicert.com`.

Note, the full path to the certificate must be given - it won't be found by MSBuild otherwise (it will look for the file with the .csproj file, where it doesn't exist), so the `git rj build` command was extended to define a new environment variable `CDIR` which is the current directory.

Issue: DOTNET-353