Git RJ: Try to fix crash on closing of a pipe
A single instance occurred:

Exception in thread Thread-7:
Traceback (most recent call last):
  File "C:\Program Files\Python39\lib\threading.py", line 954, in _bootstrap_inner
    self.run()
  File "C:\Users\jcurl\Documents\Programming\rjcp.base\configmgr\git-rj.py", line 41, in run
    for line in iter(self.pipeReader.readline, ''):
OSError: [Errno 9] Bad file descriptor

There was no output, not clear why it failed. We'll ignore the error and close it. It says that a pipe wasn't created, but no error was returned by the OS.

It's suspected that this occurred because the pipe was closed (which is OK) and no output was printed because Python likes to cache output. Worst case the process didn't work, hopefully it's caught in the process output.