MSBuildTasks: Add infrastructure to get GIT info
Provide classes and a framework to get source control information. This first implementation only supports GIT. This will be used later by a new MSBuild task to be part of the build system.

The RevisionControl task is intended to get important revision control information from the current path. A project can use this information to update metadata for the binaries it builds.

Detailed documentation and design information provided in `docs/design/revision.md`.

Issue: DOTNET-375