MSBuildTasks: AsyncValue Deadlock can occur
Under some conditions where waiting on a AsyncValue and setting it at the same time might result in a race condition and a deadlock.

Updated access to the running state to be queried via a volatile read (although this doesn't fix the problem). The problem is fixed by properly setting the state of the AsyncManualResetEvent to the "done" state when being initialized.

Issue: DOTNET-385