MSBuildTasks: Implement RevisionControl Task and SemVer Task
The RevisionControl task can help a project define the label that should be applied to a project, get the commit information and branch information, which can then be used to determine how to modify metadata for the library.

It is not as full featured as other projects, such as GitTools.GitVersion (https://github.com/GitTools/GitVersion). It uses the native GIT installed on your operating system (instead of using a precompiled library).

One of the known issues when testing is that caching sometimes causes problems as the `dotnet` binary remains running in the background, so changes to the revision control system aren't always noticed.

Issue: DOTNET-375