MSBuildTasks: Add CachedMode
Don't cache the GIT settings between builds in memory, in case that the host runs for longer than the build.

A situation can occur when running Visual Studio Code, and compiling on the PowerShell command line. Checking in new changes would not be reflected when recompiling, unless Cached="disabled", or Visual Studio Code was closed.

Another side effect of this is when building the RJCP.MSBuildTasks.csproj in debug mode, there would be an error because the file is in use. Close Visual Studio Code and the file can be built.

To reproduce:

0. Start from a clean directory (e.g. `git rj clean`).
1. Compile the debug version `dotnet build`. This is needed for compiling the release version. The debug version is what is actually tested here.
2. Modify the file RJCP.MSBuildTasks.targets so that Cached="enabled"
3. Compile the release version `dotnet build -c release`. This loads the library. It appears a few instances of `dotnet` start and remain. You'll get a dump of properties saying the commit and that it is dirty.
4. Commit the change enabling caching to git.
5. Recompile the release version `dotnet build -c release` from PowerShell. The new commit is not used. This is the bug that is because caching is enabled
6. Modify the file RJCP.MSBuildTasks.targets so that Cached="disabled" and recompile. It now uses the new commit

Issue: DOTNET-375