MSBuildTasks: Add RevisionControlClearCache
Sometimes when building, the dotnet host which contains the method that runs the RevisionControl task doesn't end, so any cached results from the ISourceProvider from the previous run remains in memory. This means when modifying the revision control system and rebuilding, outdated and incorrect results are returned.

This patch set clears the source provider cache at the beginning of the build, using inner builds with MSBuild to run only once at the start. That means when running multiple times for the same project, the results are cached.

Issue: DOTNET-375