MSBuildTasks: Update all libraries to build with GIT version information
When updating all projects, the build instructions for RJCP.MSBuildTasks no longer work exactly as expected, because the test library depends on `framework/nunitextensions`. This now imports RJCP.MSBuildTasks, which will of course fail, as it doesn't exist. So the NuGet package needs to be created first, provided at least in a local feed, then it will work.

Issue: DOTNET-375