MSBuildTask: Fix race condition executing processes (found on Linux)
Handle the case on Linux, that sometimes a Process may call the Exit event during, or immediately after Process.Start() is called. This causes a race condition, where if ProcessWorker.Dispose() is called in the event, then ProcessWorker.Start() may fail with some weird error codes (such as StandardError has not been redirected, even when it has been set in the constructor).

The solution is to handle the Dispose() method that if we haven't finished initialization, we mark a flag that the Start() should dispose the object for us at the end to avoid memory leaks. Else Dispose() can do the work itself if completed.

Issue: DOTNET-391