MSBuildTasks: Allow Build Task to work with .NET Framework and .NET Core
Visual Studio will complain about not loading projects dependent on a .NET Standard 2.1 build task, of which this project v0.2.0 was built (works fine with .NET Core).

The `RJCP.MSBuildTasks.dll` must be available as a .NET Framework component as well as .NET Core. This requires the project to provide two versions of the build task. The `RJCP.MSBUildTasks.targets` file now checks for the flavour of MSBuild, and chooses the correct assembly based on that.

Changes in code is downgrading language support from C# 8.0 to C# 7.0 (Enums and Index operators) to compile .NET 4.8,

Issue: DOTNET-392