MSBuildTasks: Errors from GIT are now more obvious
When an error is raised from GIT, it was difficult to determine the root cause from the build logs. Now the command that was executed, the working directory, the exit code and the output of the command are reported to the user.

To make the implementation easier, a new specific exception was created, that can raise the data at the point the exception is found, which contains the precise command and its results that lead to the error.

Issue: DOTNET-393