MSBuildTasks: Set the SourceRevisionId to be -dirty if needed
Prior to this commit, the RevisionControlDirty only controlled if we could make a release version or not. That couldn't be distinguished between the revision control file system being dirty, or if it is checked in but changed since the last label, or just built in Debug mode.

Now if the revision control system is dirty, the keyword '-dirty' is appended to the SourceRevisionId

Issue: DOTNET-394