MSBuildTasks: Add the 'none' SourceProvider
The 'none' source provider indicates there is no revision control. The revision control tasks are still run and still modify properties.

This is especially useful when providing `<RevisionControl>auto</RevisionControl>`, as if the provider can't be determined, the assumption is now that it is `none`.

The SourceFactory CreateAsync must now by marked as `async` to ensure that code is run at the right places within the try/catch block for automatic detection.

Issue: DOTNET-388