MSBuildTasks: Removing caching from the MSBuild Task
Caching in the library doesn't work as expected as the .NET build processes run for a long period of time. Despite having a mechanism to clear the cache, and that this is often run, it doesn't work all the time.

The best case is to always run the task when needed, but then only run the task if it needs to be run for the project.

Issue: DOTNET-412