SerialPortStream: Add new project
Update and convert my SerialPortStream project into a unified collection of projects, allowing the SerialPortStream to reuse existing code for other projects as well.

The main project, RJCP.Framework.sln, doesn't contain the integration tests for the SerialPortStream, as these often fail, but shouldn't prevent a build and a release. Modifications to the project should be built in addition against the SerialPortStream.sln file in that project.

Issue: DOTNET-398