Build: Add delays when executing processes when it ends
Using these classes in another project shows often errors when a process ends (starting a new process, or that when the process ends, files and output are not yet complete indicating the process really hasn't ended yet). Adding two delays, after the wait, and after the pipe is closed, resolves these errors.

Obviously this is a workaround (delays indicate a lack of synchronization), but it's not clear why this is the root cause at this time, and preventing build failures with a minor delay is a benefit.

Issue: DOTNET-469