Log: New module that provides a framework for reading log files
The RJCP.Diagnostics.Log namespace introduces classes that make it easier to read log files in a structured way. The first type of log file that can be read are text files, that have Unix, Windows or Classic MAC file endings. These are split up into individual lines that can be analysed.

It is more than just a Text reader, the input sources could be any type of structured packet based data. So long as it is possible to read a stream, and represent a packet as a class, a decoder can be written and used by this library.

The SandCastle documentation for this library is not present in the SHFB project, until an upgrade is done that works with .NET Core. So it is currently commented out.

Issue: DOTNET-476, DOTNET-478