git-rj: git rj pull --force now does a reset
When working between machines, and performing commit --amend, need a quick way to update the repository to be the same as the remote, automatically discarding the local changes.

The existing implementation tried to do this with a force on the branch checkout, but this doesn't work. Instead, fetch the remote, and to a hard reset to the same commit as the new fetched remote.