Update .gitrjbuild for building on Linux
Building on Linux (Ubuntu 22.04) still works when Mono is installed, as well as a framework (e.g. .NET 7.0). Update the build scripts to suppress output on the console.

On Linux, the `-v=q` option doesn't work, so `--logger=\"console;verbosity=quiet\"` is required, which suppresses the output.

I didn't find out how to execute the testhost command on Linux with Mono, so suppress the frameworks that are tested to .NET Core only. Note, this part is rather "hacky" and a better solution should be found. Also note, that there are still complaints when the project uses `<TargetFramework>net48</TargetFramework>` instead of `<TargetFrameworks>...</TargetFrameworks>`. It would be worthwhile to further investigate if we can disable building/testing frameworks depending on the installed SDKs.

Issue: DOTNET-835