MSBuildTasks: Fix process start race conditions on Linux
Failures on Linux would often be seen during build while trying to access `git`. Either unexpected output, missing output or exceptions in the task due to unexpected output.

The root cause of the issue came down to the asynchronous behaviour of using the Microsoft `Process` class on Linux (which works fine on Windows). AFter doing endurance tests, the process would raise the `Exited` event before returning from the `Process.Start()` method, thus unable to capture the output.

It is not possible to workaround to get the output of the `stdout` or `stderr` on the exit event, because the properties `Process.StandardOutput` and `Process.StandardError` raise an `InvalidOperationException` if `ReadLine{Async}` is called because they've been configured for events. But the `Process.BeginOutputReadLine` or `Process.BeginErrorReadLine` haven't had time to start yet! And if they do, the process has already exited, so another kind of error is caused.

As such, a synchronous implementation model must be implemented, where we can control our own `ProcessExitEvent` when the process has exited and we've read all the output. We know the user has read all the output (the threads on which they're read is implementation defined), because we synchronise such that all events are finished first before raising `ProcessExitEvent`.

Tested on Linux by cycling this project about 1500 times with no errors, where before it would occur approximately 1 build in 10.

Issue: DOTNET-838