Environment Version: Rework .NET Runtime detection
Simplifies (and reworks API) for detecting the .NET Framework and .NET Core runtimes. Now detects Mono run-times on Linux and Windows, from version 1.x to the current version, also on Windows XP.

Detection uses the underlying file-system to look for duplicates, and so this functionality is added to `RJCP.Path`. To avoid cyclic dependencies, a new assembly `RJCP.Environment.Version` is created (as `Path` needs the platform information in `Environment`).

The `NetVersion` interface was cleaned up with some breaking changes (for the better and simpler).

To support `NetVersion`, `Path` now contains a generic `FileSystemNodeInfo` that uses Windows APIs and Glibc 6 APIs to query the underlying filesystem. This allows us to detect duplicates easily, by having essentially inode information we can compare with. So the list of Mono runtimes is only presented once to the user in available runtimes.

Similarly, to support `NetVersion`, we want to check that the binary is executable and get the architecture for that executable. So a new `FileExecutable` class is implemented, that when queried, knows how to interpret an ELF file or a Windows PE file. These implementations are reasonably generic and can be expanded upon.

Issue: DOTNET-848