Upgrade to .NET Core 6.0 (and move away from .NET Standard 2.1)
1. Major change to .NET 6.0

All references of .NET Standard are upgraded to .NET 6.0. All .NET Core dependencies are updated to .NET Core 6.0.x (DOTNET-936).

RJCP.MSBuildTasks remains on .NET Standard 2.1 to be compatible with as many projects as possible. As it's a DLL and it has no dependencies (either on it, or on other projects), leaving at .NET Standard 2.1 is the most suitable action.

The tag `RollForward` is set to `Major` for all `OutputType` of `Exe`.

2. Multi-targetting for Frameworks and Operating System

A new library, `RJCP.Core.SysCompat` is created to further support multiframework targetting, where .NET 4.x is left behind. Some wrappers and stubs are created to allow us to upgrade the framework versions, or to avoid excessive use of conditional code.

Code is now annotated with `[SupportedOSPlatform(xxx)]` to help with multitargetting different frameworks and Operating Systems (DOTNET-938). This will bring a big benefit to using the libraries with the proper annotations.

`ThrowHelper.ThrowIfNull` replaces `ArgumentNullException.ThrowIfNull` (DOTNET-945)

3. Miscellaneous Changes

All warnings with the upgrade to .NET 6.0 are removed. Serialization (DOTNET-951), Constrained Execution Regions (DOTNET-937), Code Access Security (DOTNET-943, DOTNET-940) are not supported in .NET Core, but were updated to be consistent for .NET Framework (or removed where it made no sense to keep).

Upgrade where possible to use native .NET 6.0 methods (DOTNET-948, DOTNET-944). Remove reflection (because feature isn't available in .NET Standard but now in .NET 6.0, DOTNET-952). Other minor performance benefits (DOTNET-946, DOTNET-947).

4. Style Changes

Many IDE style changes are now present (DOTNET-942, DOTNET-959), which are resolved by upgrading .NET Framework to use C# 10.0, so they don't need to be ignored in a `GlobalSuppression.cs` file and can be applied. The most important change is to use `myObj is null` instead of `myObj == null`.

5. Fixes for Linux and Mono

Reenabled testing .NET Framework libraries (`.gitrjbuild`) on Ubuntu 20.04/22.04 as it appears now to work with `dotnet-sdk-6.0` (MSBuild version 17.3.2+561848881 for .NET) and mono 6.12.0.200 (or 6.8.0.105).

Fixes for specific libraries:
* CrashReporter now passes all test cases (DOTNET-969), with flaky tests before.
* DltDump now properly tests with time zones (DOTNET-967).
* Thread.GroupTasks race condition resolved (DOTNET-970).

Issue: DOTNET-892