MSBuild: Allow overrides in the file .rjbuild
When building in strict mode, it is expected there are no changes to the repository since the label was applied. This is useful to ensure that a release build is not applied if it is not intended (i.e. the file changed).

In some cases it is desirable to override this behaviour (because there was a change to a file, like the .csproj file) that doesn't affect the build (e.g. it pulls in analyzers and the like but doesn't affect compilation). In this case, it should be possible to provide an override.

Create a file in the external repository (not the repo that is being build, but in this case the base repository), called `.rjbuild`. In it, create with the contents:

[git-overrides]
tag = <hash>

The hash must be the hash of the HEAD that is being built. If the HEAD of the repository matches the hash in the configuration file for the tag that is being checked, then it is overridden and the differences are allowed. On the next commit, it will be ignored (and must be changed again) because the hash of the HEAD has changed. So for the next release, need to check again and update (or remove) as necessary.

Issue: DOTNET-1029