RJCP Documentation: Fix MSHC System Path with 2025.03.22
When building documentation using SHFB, the library `RJCP.Core.SysCompat` provides some classes in the namespace `N:System`. Sandcastle reports a benign warning:

`SHFB: Warning BE0066: CopyFromIndexComponent: Entries for the key 'N:System' occur in both ...`

But this can be ignored. What actually causes a problem is that the file describing the namespace `N:System` is put in the correct location when imported into MS Help Viewer 2.3, but then all .NET 4.0 Framework documentation is moved to this new folder, not where it should be.

This plugin changes the MSHC build documentation meta data that the topic name no longer overlaps.

It is a very simple plugin, and only intended for this project, not for general use, because there is no configuration.

Issue: DOTNET-1067, https://github.com/EWSoftware/SHFB/issues/1131