RjBuild: Ignore out of service SDKs
We might still want to build for older SDKs, hot fixes, etc. There's no reason that we have to go back to an older SDK to do this, but we do want to ignore the SDK.

Warnings are still kept, and will still be visible in Visual Studio. Building from the command line with `git rj build` will now ignore the warning.

Issue: DOTNET-1076